修改 vs.net 工程文件，修复一处编译警告

+++ fibjs/include/MongoID.h
@@ -43,7 +43,7 @@ public:
 		std::string str;
 
 		toString(str);
-		retVal = v8::String::New(str.c_str(), str.length());
+		retVal = v8::String::New(str.c_str(), (int)str.length());
 
 		return 0;
 	}
